requirements.txt
aldryn-apphooks-config==0.5.2
aldryn-background-image==0.1.1
aldryn-boilerplates==0.8.0
aldryn-bootstrap3==1.3.0
aldryn-categories==1.2.0
aldryn-common==1.0.5
aldryn-people==2.2.0
aldryn-redirects==1.0.0
aldryn-search==1.0.1
aldryn-translation-tools==0.3.0
argcomplete==1.11.1
Babel==2.7.0
backport-collections==0.1
boto3==1.14.12
botocore==1.17.12
certifi==2019.6.16
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
cmsplugin-filer==1.1.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.17
django-admin-sortable==2.0.18
django-allauth==0.39.1
django-amp-tools==0.2.2
django-appconf==1.0.3
django-appdata==0.2.1
django-classy-tags==0.9.0
django-cms==3.6.0
django-filer==1.5.0
django-formtools==2.1
django-haystack==2.8.1
django-hero-slider==1.3.1
django-hvad==1.8.0
django-js-asset==1.2.2
django-libs==2.0.2
django-meta==1.5.2
django-meta-mixin==0.3.0
django-mptt==0.10.0
django-parler==2.0
django-phonenumber-field==1.3.0
django-polymorphic==2.0.3
django-recaptcha==2.0.6
django-reset-migrations==0.3.1
django-rss-plugin==0.1.0
django-sekizai==1.0.0
django-simple-math-captcha==1.0.9
django-sortedm2m==1.5.0
django-spurl==0.6.6
django-standard-form==1.1.1
django-storages==1.7.1
django-taggit==0.23.0
django-taggit-autosuggest==0.3.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-apphook-setup==0.3.0
djangocms-attributes-field==1.1.0
djangocms-blog==1.0.0b1
djangocms-column==1.9.0
djangocms-custommenu==1.0.2
djangocms-file==2.3.0
djangocms-flexslider==1.0.0
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.1.1
djangocms-link==2.5.0
djangocms-picture==2.1.3
djangocms-redirect==0.3.1
djangocms-slick-slider==0.5.1
djangocms-slider-easy==0.2.6
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework==3.9.4
docutils==0.14
durationpy==0.5
easy-thumbnails==2.6
elasticsearch==7.0.2
feedparser==5.2.1
fields==5.0.0
future==0.18.2
hjson==3.0.1
html5lib==1.0.1
idna==2.8
image==1.5.27
importlib-metadata==1.7.0
jmespath==0.10.0
jsonfield==2.0.2
kappa==0.6.0
lxml==4.4.0
mysqlclient==1.4.6
oauthlib==3.1.0
olefile==0.46
phonenumbers==8.10.15
phonenumberslite==8.10.15
Pillow==6.1.0
pip-tools==5.2.1
placebo==0.9.0
pysolr==3.8.1
python-dateutil==2.6.1
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.12.0
South==1.0.2
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.1
tqdm==4.47.0
troposphere==2.6.1
tzlocal==2.0.0
Unidecode==1.0.23
urllib3==1.25.3
URLObject==2.4.3
virtualenv==16.7.2
webencodings==0.5.1
Werkzeug==0.16.1
Whoosh==2.7.4
wsgi-request-logger==0.4.6
YURL==1.0.0
zappa==0.51.0
zipp==3.1.0